Kinoite: Add plasma-welcome-fedora
This package is explicitely not pulled via the comps group in order to
not be included by default in Fedora derivatives. We don't have a way to
signal that for Kinoite so include it here. Derivatives will have to
remove / disable the package as needed.

See: https://pagure.io/fedora-kde/SIG/issue/295